requirements_app.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyotp>=2.9.0
qrcode[pil]>=7.4.2
python-multipart>=0.0.6
anthropic>=0.34.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
websockets>=12.0
